Added login-based access to database

+++ web/update.php
@@ -23,14 +23,38 @@
 
   foreach ($queue['updates'] as $name => $collection) {
     $collectionId = $db->getCollectionId($name);
+    $canUpdate = $db->canUpdate($name);
+
+    if (!$canUpdate) {
+      continue;
+    }
+
 //print "Collection id for $name: $collectionId\n";
+
+    // Check to see if the collection is clean
+    $clean = array();
     foreach ($collection as $key => $value) {
-      $db->updateItem($collectionId, $key, $value);
+      try {
+        $encoded = json_encode($value);
+        $clean[$key] = $encoded;
+      } catch (Exception $inner) {
+        print "{\"result\":9091435}";
+        return;
+      }
+    }
+
+    foreach ($clean as $key => $encoded) {
+      $db->updateItem($collectionId, $key, $encoded);
     }
   }
 
   foreach ($queue['deletes'] as $name => $collection) {
     $collectionId = $db->getCollectionId($name);
+    $canDelete = $db->canDelete($name);
+
+    if (!$canDelete) {
+      continue;
+    }
 
     foreach ($collection as $key => $value) {
       $db->deleteItem($collectionId, $key);
